moved extras inside src and updated import paths
