--- extras/.eslintrc.json
@@ -1,5 +1,0 @@
-{
-    "rules": {
-        "import/no-unresolved": "off"
-    }
-}
