--- extras/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "playcanvas-extras",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "type": "module",
-  "author": "",
-  "license": "MIT",
-  "dependencies": {
-    "playcanvas": "file:.."
-  }
-}
