+++ src/extras/mini-stats/gpu-timer.js
