+++ src/extras/mini-stats/mini-stats.js
@@ -1,10 +1,8 @@
-import {
-    ADDRESS_REPEAT,
-    FILTER_NEAREST,
-    LAYERID_UI,
-    math,
-    Texture
-} from 'playcanvas';
+import { math } from '../../core/math/math.js';
+import { Texture } from '../../platform/graphics/texture.js';
+import { ADDRESS_REPEAT, FILTER_NEAREST } from '../../platform/graphics/constants.js';
+import { LAYERID_UI } from '../../scene/constants.js';
+
 import { CpuTimer } from './cpu-timer.js';
 import { GpuTimer } from './gpu-timer.js';
 import { StatsTimer } from './stats-timer.js';
@@ -21,7 +19,7 @@ class MiniStats {
     /**
      * Create a new MiniStats instance.
      *
-     * @param {import('playcanvas').AppBase} app - The application.
+     * @param {import('../../framework/app-base.js').AppBase} app - The application.
      * @param {object} [options] - Options for the MiniStats instance.
      * @param {object[]} [options.sizes] - Sizes of area to render individual graphs in and spacing
      * between individual graphs.
@@ -262,8 +260,8 @@ class MiniStats {
     /**
      * Create the graphs requested by the user and add them to the MiniStats instance.
      *
-     * @param {import('playcanvas').AppBase} app - The application.
-     * @param {import('playcanvas').GraphicsDevice} device - The graphics device.
+     * @param {import('../../framework/app-base.js').AppBase} app - The application.
+     * @param {import('../../platform/graphics/graphics-device.js').GraphicsDevice} device - The graphics device.
      * @param {object} options - Options for the MiniStats instance.
      * @private
      */
