+++ src/extras/render-passes/render-pass-downsample.js
@@ -1,12 +1,10 @@
-import {
-    RenderPassShaderQuad
-} from "playcanvas";
+import { RenderPassShaderQuad } from '../../scene/graphics/render-pass-shader-quad.js';
 
 class RenderPassDownsample extends RenderPassShaderQuad {
     constructor(device, sourceTexture) {
         super(device);
         this.sourceTexture = sourceTexture;
-        this.shader = this.createQuadShader('DownSampleShader', `
+        this.shader = this.createQuadShader('DownSampleShader', /* glsl */`
 
             uniform sampler2D sourceTexture;
             uniform vec2 sourceInvResolution;
