+++ src/extras/render-passes/render-pass-prepass.js
@@ -1,14 +1,16 @@
 import {
     LAYERID_DEPTH,
-    SHADER_PREPASS_VELOCITY,
+    SHADER_PREPASS_VELOCITY
+} from '../../scene/constants.js';
+import {
     FILTER_NEAREST,
     PIXELFORMAT_RGBA32F,
     PIXELFORMAT_RGBA16F,
-    ADDRESS_CLAMP_TO_EDGE,
-    Texture,
-    RenderPass,
-    RenderTarget
-} from "playcanvas";
+    ADDRESS_CLAMP_TO_EDGE
+} from '../../platform/graphics/constants.js';
+import { Texture } from '../../platform/graphics/texture.js';
+import { RenderPass } from '../../platform/graphics/render-pass.js';
+import { RenderTarget } from '../../platform/graphics/render-target.js';
 
 const tempMeshInstances = [];
 
@@ -26,7 +28,7 @@ const VELOCITY_UNIFORM_NAME = 'uSceneVelocityMap';
  * @ignore
  */
 class RenderPassPrepass extends RenderPass {
-    /** @type {import("playcanvas").BindGroup[]} */
+    /** @type {import('../../platform/graphics/bind-group.js').BindGroup[]} */
     viewBindGroups = [];
 
     /** @type {Texture} */
