+++ src/extras/render-passes/render-pass-upsample.js
@@ -1,13 +1,11 @@
-import {
-    RenderPassShaderQuad
-} from "playcanvas";
+import { RenderPassShaderQuad } from '../../scene/graphics/render-pass-shader-quad.js';
 
 class RenderPassUpsample extends RenderPassShaderQuad {
     constructor(device, sourceTexture) {
         super(device);
         this.sourceTexture = sourceTexture;
 
-        this.shader = this.createQuadShader('UpSampleShader', `
+        this.shader = this.createQuadShader('UpSampleShader', /* glsl */`
 
             uniform sampler2D sourceTexture;
             uniform vec2 sourceInvResolution;
